+++ fvdb/viz/__init__.py
@@ -4,6 +4,7 @@
 from ._camera_view import CamerasView
 from ._fog_volume_view import FogVolumeView
 from ._gaussian_splat_3d_view import GaussianSplat3dView
+from ._gaussian_splat_view_data import GaussianSplatViewData
 from ._image_view import ImageView
 from ._level_set_view import LevelSetView
 from ._point_cloud_view import PointCloudView
@@ -17,6 +18,7 @@ __all__ = [
     "wait_for_interrupt",
     "FogVolumeView",
     "GaussianSplat3dView",
+    "GaussianSplatViewData",
     "CamerasView",
     "ImageView",
     "LevelSetView",